+++ src/DotNext.Metaprogramming/Runtime/CompilerServices/IAsyncStateMachine.cs
@@ -34,11 +34,19 @@ internal interface IAsyncStateMachine<out TState> : IAsyncStateMachine
         where TException : Exception;
 }
 
+internal interface IAsyncStateMachine<TState, out TValueTask, TSelf> : IAsyncStateMachine<TState>
+    where TState : struct
+    where TValueTask : struct, IEquatable<TValueTask>
+    where TSelf : struct, IAsyncStateMachine<TState, TValueTask, TSelf>
+{
+    static abstract TValueTask Start(Transition<TState, TSelf> transition, TState initialState);
+}
+
 /// <summary>
 /// Represents body of async method in the form of state machine transitions.
 /// </summary>
 /// <typeparam name="TState">The type of async method state.</typeparam>
 /// <typeparam name="TMachine">The implementation of async state machine.</typeparam>
-/// <param name="stateMachine">Asyncronous state machine.</param>
+/// <param name="stateMachine">Asynchronous state machine.</param>
 internal delegate void Transition<out TState, TMachine>(ref TMachine stateMachine)
     where TMachine : struct, IAsyncStateMachine<TState>;