+++ src/DotNext.Metaprogramming/Threading/Tasks/CompletedTask.cs
@@ -1,6 +1,4 @@
-﻿using System.Diagnostics.CodeAnalysis;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
+﻿using System.Runtime.InteropServices;
 
 namespace DotNext.Threading.Tasks;
 
@@ -37,41 +35,31 @@ internal readonly struct CompletedTask(Exception failure)
 [StructLayout(LayoutKind.Auto)]
 internal readonly struct CompletedTask<T>
 {
-    private readonly Exception? failure;
-
-    [AllowNull]
-    private readonly T result;
+    private readonly Result<T> result;
 
     /// <summary>
     /// Creates task that has completed with a specified exception.
     /// </summary>
     /// <param name="failure">The exception with which to complete the task.</param>
-    public CompletedTask(Exception failure) => this.failure = failure;
+    public CompletedTask(Exception failure) => result = new(failure);
 
     /// <summary>
     /// Creates task that has completed successfully with a specified result.
     /// </summary>
     /// <param name="result">The task result.</param>
-    public CompletedTask(T result) => this.result = result;
+    public CompletedTask(T result) => this.result = new(result);
 
     /// <summary>
     /// Obtains <see cref="Task{TResult}"/> completed synchronously.
     /// </summary>
     /// <param name="task">Completed task.</param>
     public static implicit operator Task<T>(CompletedTask<T> task)
-        => task.failure is null ? Task.FromResult(task.result) : Task.FromException<T>(task.failure);
+        => task.result.AsTask().AsTask();
 
     /// <summary>
     /// Obtains <see cref="ValueTask{TResult}"/> completed synchronously.
     /// </summary>
     /// <param name="task">Completed task.</param>
     public static implicit operator ValueTask<T>(CompletedTask<T> task)
-    {
-        var builder = AsyncValueTaskMethodBuilder<T>.Create();
-        if (task.failure is null)
-            builder.SetResult(task.result);
-        else
-            builder.SetException(task.failure);
-        return builder.Task;
-    }
+        => task.result.AsTask();
 }