+++ src/DotNext.Tests/Net/Http/HttpEndPointTests.cs
@@ -1,6 +1,5 @@
 using System.Net.Sockets;
 using System.Text;
-using Microsoft.AspNetCore.Components.Forms;
 
 namespace DotNext.Net.Http;
 
@@ -107,5 +106,14 @@ public sealed class HttpEndPointTests : Test
         static bool TryParse<T>(string input, out T result)
             where T : IParsable<T>
             => T.TryParse(input, provider: null, out result);
+    }
+
+    [Theory]
+    [InlineData("192.168.0.1", AddressFamily.InterNetwork)]
+    [InlineData("2001:0db8:0000:0000:0000:8a2e:0370:7334", AddressFamily.InterNetworkV6)]
+    [InlineData("host", AddressFamily.Unspecified)]
+    public static void AutoDetectAddressFamily(string hostName, AddressFamily expected)
+    {
+        Equal(expected, new HttpEndPoint(hostName, 8080, false).AddressFamily);
     }
 }